Fix #636 by killing preview terminal
