+++ src/guake/guake_notebook.py
@@ -5,11 +5,8 @@ from __future__ import print_function
 import logging
 import os
 import posix
-import signal
 
 from gtk import Notebook
-from thread import start_new_thread
-from time import sleep
 
 log = logging.getLogger(__name__)
 
@@ -75,48 +72,13 @@ class GuakeNotebook(Notebook):
 
     def delete_tab(self, pagepos, kill=True):
         for terminal in self.get_terminals_for_tab(pagepos):
-            pid = terminal.get_pid()
             if kill:
-                start_new_thread(self.delete_shell, (pid,))
+                terminal.kill()
 
             terminal.destroy()
 
         self.remove_page(pagepos)
         self.term_list.pop(pagepos)
-
-    def delete_shell(self, pid):
-        """This function will kill the shell on a tab, trying to send
-        a sigterm and if it doesn't work, a sigkill. Between these two
-        signals, we have a timeout of 3 seconds, so is recommended to
-        call this in another thread. This doesn't change any thing in
-        UI, so you can use python's start_new_thread.
-        """
-        try:
-            os.kill(pid, signal.SIGHUP)
-        except OSError:
-            pass
-        num_tries = 30
-
-        while num_tries > 0:
-            try:
-                # Try to wait for the pid to be closed. If it does not
-                # exist anymore, an OSError is raised and we can
-                # safely ignore it.
-                if os.waitpid(pid, os.WNOHANG)[0] != 0:
-                    break
-            except OSError:
-                break
-            sleep(0.1)
-            num_tries -= 1
-
-        if num_tries == 0:
-            try:
-                os.kill(pid, signal.SIGKILL)
-                os.waitpid(pid, 0)
-            except OSError:
-                # if this part of code was reached, means that SIGTERM
-                # did the work and SIGKILL wasnt needed.
-                pass
 
     def append_tab(self, terminal):
         self.term_list.append(terminal)
