+++ src/guake/terminal.py
@@ -28,6 +28,7 @@ import logging
 import os
 import pango
 import re
+import signal
 import subprocess
 import uuid
 import vte
@@ -35,6 +36,8 @@ import vte
 
 from guake.common import clamp
 from guake.globals import KEY
+from thread import start_new_thread
+from time import sleep
 
 log = logging.getLogger(__name__)
 
@@ -268,6 +271,44 @@ class GuakeTerminal(vte.Terminal):
 
     def decrease_font_size(self):
         self.font_scale -= 1
+
+    def kill(self):
+        pid = self.get_pid()
+        start_new_thread(self.delete_shell, (pid,))
+
+    def delete_shell(self, pid):
+        """This function will kill the shell on a tab, trying to send
+        a sigterm and if it doesn't work, a sigkill. Between these two
+        signals, we have a timeout of 3 seconds, so is recommended to
+        call this in another thread. This doesn't change any thing in
+        UI, so you can use python's start_new_thread.
+        """
+        try:
+            os.kill(pid, signal.SIGHUP)
+        except OSError:
+            pass
+        num_tries = 30
+
+        while num_tries > 0:
+            try:
+                # Try to wait for the pid to be closed. If it does not
+                # exist anymore, an OSError is raised and we can
+                # safely ignore it.
+                if os.waitpid(pid, os.WNOHANG)[0] != 0:
+                    break
+            except OSError:
+                break
+            sleep(0.1)
+            num_tries -= 1
+
+        if num_tries == 0:
+            try:
+                os.kill(pid, signal.SIGKILL)
+                os.waitpid(pid, 0)
+            except OSError:
+                # if this part of code was reached, means that SIGTERM
+                # did the work and SIGKILL wasnt needed.
+                pass
 
 
 class GuakeTerminalBox(gtk.HBox):
